Bands

+++ database/migrations/2018_11_11_211017_create_photos_table.php
@@ -12,8 +12,8 @@ class CreatePhotosTable extends Migration {
 	 */
 	public function up() {
 		Schema::create('photos', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('kitchen_id')->unsigned();
+			$table->bigIncrements('id');
+			$table->bigInteger('kitchen_id')->unsigned();
 			$table->string('file');
 			$table->timestamps();
