+++ database/migrations/2018_11_20_230939_create_products_table.php
@@ -12,8 +12,8 @@ class CreateProductsTable extends Migration {
 	 */
 	public function up() {
 		Schema::create('products', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('application_id')->unsigned();
+			$table->bigIncrements('id');
+			$table->bigInteger('application_id')->unsigned();
 			$table->string('name');
 			$table->decimal('price');
 			$table->string('category');
