+++ database/migrations/2018_12_10_211753_create_invoice_payments_table.php
@@ -12,10 +12,10 @@ class CreateInvoicePaymentsTable extends Migration {
 	 */
 	public function up() {
 		Schema::create('invoice_payments', function (Blueprint $table) {
-			$table->increments('id');
+			$table->bigIncrements('id');
 			$table->decimal('amount');
 			$table->date('date');
-			$table->integer('invoice_id')->unsigned();
+			$table->bigInteger('invoice_id')->unsigned();
 
 			$table->timestamps();
