+++ database/migrations/2019_01_21_201052_create_worker_workplace_table.php
@@ -12,9 +12,9 @@ class CreateWorkerWorkplaceTable extends Migration {
 	 */
 	public function up() {
 		Schema::create('worker_workplace', function (Blueprint $table) {
-			$table->increments('id');
-			$table->integer('worker_id')->unsigned();
-			$table->integer('workplace_id')->unsigned();
+			$table->bigIncrements('id');
+			$table->bigInteger('worker_id')->unsigned();
+			$table->bigInteger('workplace_id')->unsigned();
 
 			$table->timestamps();
